Refactor HttpPathHandler/HttpEndpointHandler hierarchy
Separate path handling and endpoint handling logic, as they
fundamentally do different thing (one dispatches, one handles).

In addition, the current design reuses the same dispatcher instance for
all requests to the same endpoint. However, since 8fb14893 the
abstract dispatcher base class contains state, which may cause data race
when there is a high request throughput.

This commit forces all requests to first dispatch by a path handler
(which is stateless), then dispatch to a new instance of the endpoint
handler. This change results in the introduction of LoginHandler and
LogoutHandler as path handlers.

Utility methods which used to reside in HttpPathHandler is now moved to
HttpUtils.