[DatabaseConnectionV2] Fix timeout handling
Today I learned that InnoDB's lock timeout and the transaction timeout
required by this project are two vastly different things, and should be
treated differently. As such, the timeouts are now encapsulated within
the connection pool, only intending for tests to manipulate them
on-the-fly.

In addition, to reduce the time required to run test cases, these
timeouts are configurable. The time required to execute Transaction*
tests are now vastly reduced.

JMeter scenarios are not tested yet.